added storing private messages. Changed tests.
Private messages are now stored with messageService, so even test users now have to be stored to database. That requires moving online user storing to userService, so it can be handled globally.
initial-user is now actually created and not just a mock user in memory.
tests are changed to work with initial-user changes.
All tests work.
Saving messages is not yet awaited, as socket io initialization has to be changed to async.