Fix to handle multiple connections from user
Connection id is now also saved with connected user, but only distinct users are returned to client.
This prevents a bug where same user would log in on two clients and when logging out the second client, the user would be marked as 'false offline' while still being in the chat.