Changed message and event Redis storing method
Instead of storing messages in a list that has a shared TTL, now storing them under separate keys.
This allows single messages to expire even when new messages are constantly stored.

When using Redis this way, all existing and matching keys need to be scanned first and only then fetched.